+++ src/main/java/pl/tw/ksiegowosc/dto/IssueAllegroInvoiceResponse.java
@@ -1,0 +1,7 @@
+package pl.tw.ksiegowosc.dto;
+
+public record IssueAllegroInvoiceResponse(
+        String invoiceNo,
+        String meritInvoiceId
+) {
+}