+++ src/main/java/pl/tw/ksiegowosc/dto/MeritCustomersRequest.java
@@ -5,6 +5,10 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public record MeritCustomersRequest(
-        @JsonProperty("Name") String name
+        @JsonProperty("Name") String name,
+        @JsonProperty("VatRegNo") String vatRegNo
 ) {
+    public MeritCustomersRequest(String name) {
+        this(name, null);
+    }
 }