+++ src/main/java/pl/tw/ksiegowosc/dto/allegro/AllegroInvoice.java
@@ -3,7 +3,7 @@ package pl.tw.ksiegowosc.dto.allegro;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-public record AllegroBuyer(
-        String login,
-        String email) {
+public record AllegroInvoice(
+        Boolean required,
+        AllegroInvoiceAddress address) {
 }