+++ src/main/java/pl/tw/ksiegowosc/dto/allegro/AllegroInvoiceAddress.java
@@ -1,0 +1,13 @@
+package pl.tw.ksiegowosc.dto.allegro;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+public record AllegroInvoiceAddress(
+        String street,
+        String city,
+        String zipCode,
+        String countryCode,
+        AllegroInvoiceCompany company,
+        AllegroNaturalPerson naturalPerson) {
+}