+++ src/main/java/pl/tw/ksiegowosc/dto/allegro/AllegroInvoiceCompany.java
@@ -1,9 +1,12 @@
 package pl.tw.ksiegowosc.dto.allegro;
 
+import java.util.List;
+
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-public record AllegroBuyer(
-        String login,
-        String email) {
+public record AllegroInvoiceCompany(
+        String name,
+        String taxId,
+        List<AllegroTaxId> ids) {
 }